# indexer/services/comlink/public/swagger.json
{
  "components": {
...
    "schemas": {
...
      "IsoString": {
        "type": "string"
      },
...
      "CandleResolution": {
        "enum": [
          "1MIN",
          "5MINS",
          "15MINS",
          "30MINS",
          "1HOUR",
          "4HOURS",
          "1DAY"
        ],
        "type": "string"
      },
      "CandleResponseObject": {
        "properties": {
          "startedAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "ticker": {
            "type": "string"
          },
          "resolution": {
            "$ref": "#/components/schemas/CandleResolution"
          },
          "low": {
            "type": "string"
          },
          "high": {
            "type": "string"
          },
          "open": {
            "type": "string"
          },
          "close": {
            "type": "string"
          },
          "baseTokenVolume": {
            "type": "string"
          },
          "usdVolume": {
            "type": "string"
          },
          "trades": {
            "type": "number",
            "format": "double"
          },
          "startingOpenInterest": {
            "type": "string"
          },
          "orderbookMidPriceOpen": {
            "type": "string",
            "nullable": true
          },
          "orderbookMidPriceClose": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "startedAt",
          "ticker",
          "resolution",
          "low",
          "high",
          "open",
          "close",
          "baseTokenVolume",
          "usdVolume",
          "trades",
          "startingOpenInterest",
          "id"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "CandleResponse": {
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/CandleResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "candles"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Indexer API",
    "version": "v1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/candles/perpetualMarkets/{ticker}": {
      "get": {
        "operationId": "GetCandles",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resolution",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CandleResolution"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "fromISO",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "toISO",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "",
      "description": "Public Testnet"
    }
  ]
}